backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==4.2.27
django-axes==8.3.1
django-cors-headers==4.6.0
django-debug-toolbar==6.2.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
redis==7.1.1
setuptools==82.0.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.46.3
whitenoise==6.11.0

google-genai
